requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
babel==2.17.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
exceptiongroup==1.3.1
fastapi==0.99.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
h11==0.16.0
htmldate==1.9.4
idna==3.11
jieba3k==0.35.1
joblib==1.5.3
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
newspaper3k==0.2.8
nltk==3.9.2
pillow==11.3.0
pydantic==1.10.26
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
starlette==0.27.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
